pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "kemomctr"
version = "0.1.0"
description = "Fast AI-powered Minecraft Mod localization tool using Gemini 3.0 Flash."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT License"}
authors = [
    {name = "kemo", email = "tennis611h14+kemomctr@gmail.com"}
]
dependencies = [
    "google-genai>=1.0.0"
]

[project.scripts]
kemomctr = "kemomctr.cli:main"